+++ classes/PdfModule.php
@@ -1,9 +1,32 @@
 <?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2017 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
 
 namespace HeimrichHannot\Share;
 
-interface PdfModuleInterface
+
+abstract class PdfModule
 {
+
+    private $module;
+
+    /**
+     * PdfModule constructor
+     *
+     *
+     */
+    public function __construct()
+    {
+
+    }
+
     /**
      * Display PDF inline or not (e.g. download it)
      *
@@ -11,7 +34,10 @@ interface PdfModuleInterface
      *
      * @return mixed true/false if supported, null if not supported
      */
-    public function setInline($inline);
+    public function setInline($inline)
+    {
+        return null;
+    }
 
     /**
      * Set pdf font size
@@ -20,7 +46,10 @@ interface PdfModuleInterface
      *
      * @return mixed int fontsize or null if not supported
      */
-    public function setFontSize($size);
+    public function setFontSize($size)
+    {
+        return null;
+    }
 
     /**
      * Set the pdf file name without file extension.
@@ -29,7 +58,10 @@ interface PdfModuleInterface
      *
      * @return mixed string filename or null if not supported
      */
-    public function setFileName($name);
+    public function setFileName($name)
+    {
+        return null;
+    }
 
     /**
      * Set login information, if pdf reader cannot use cookies from browser session.
@@ -39,19 +71,18 @@ interface PdfModuleInterface
      *
      * @return mixed user if success, null if not supported
      */
-    public function setLoginInformation($user = '', $password = '');
+    public function setLoginInformation($user = '', $password = '')
+    {
+        return null;
+    }
 
     /**
-     * Add Html content to render
+     * Render the PDF file and output it
      *
-     * @param $content string
-     *
-     * @return mixed true if added successfully, null if not supported
      */
-    public function addHtmlContent($content);
+    public function compile ()
+    {
 
-    /**
-     * Render the pdf and output it
-     */
-    public function compile();
+    }
+
 }